requirements.txt
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
Django==2.2.1
django-extensions==2.1.6
django-rest-swagger==2.2.0
djangorestframework==3.9.3
djangorestframework-stubs==0.4.2
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
matlab==0.1
matplotlib==3.0.3
numpy==1.16.3
openapi-codec==1.3.2
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
requests==2.21.0
scipy==1.2.1
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.24.3
